COMP: change target of file system completion for path string literals
